+++ app/routes/userRouts.ts
@@ -1,5 +1,5 @@
 export const userRouts = {
-  profile: '/users/profile',
+  profile: '/users/profile/1',
   avatar: '/users/profile/avatar',
 
   home: '/home',
